+++ backend/src/test/java/com/benine/backend/http/CameraIrisHandlerTest.java
@@ -1,9 +1,7 @@
-package com.benine.backend.http.jetty;
+package com.benine.backend.http;
 
 import com.benine.backend.camera.CameraConnectionException;
 import com.benine.backend.camera.IrisCamera;
-import com.benine.backend.http.CameraIrisHandler;
-import com.benine.backend.http.CameraRequestHandler;
 import org.eclipse.jetty.util.MultiMap;
 import org.junit.Before;
 import org.junit.Test;
