+++ backend/src/test/java/com/benine/backend/http/PresetsPresetCreationHandlerTest.java
@@ -1,11 +1,9 @@
-package com.benine.backend.http.jetty;
+package com.benine.backend.http;
 
 import com.benine.backend.Preset;
 import com.benine.backend.camera.CameraConnectionException;
 import com.benine.backend.camera.Position;
 import com.benine.backend.camera.ipcameracontrol.IPCamera;
-import com.benine.backend.http.PresetsCreatePresetHandler;
-import com.benine.backend.http.RequestHandler;
 import com.benine.backend.video.MJPEGStreamReader;
 import com.benine.backend.video.Stream;
 import com.benine.backend.video.StreamNotAvailableException;
