+++ backend/src/test/java/com/benine/backend/http/RequestHandlerTest.java
@@ -1,4 +1,4 @@
-package com.benine.backend.http.jetty;
+package com.benine.backend.http;
 
 import com.benine.backend.Logger;
 import com.benine.backend.PresetController;
